Fix for Bug 1398 (ifSpeed converted to bigint)

+++ etc/create.sql
@@ -293,7 +293,7 @@ create table snmpInterface (
 	snmpIfDescr		varchar(256),
 	snmpIfType		integer,
 	snmpIfName		varchar(32),
-	snmpIfSpeed		integer,
+	snmpIfSpeed		bigint,
 	snmpIfAdminStatus	integer,
 	snmpIfOperStatus	integer,
 	snmpIfAlias		varchar(256),
